correção de bug

+++ Entities/api.py
@@ -1,15 +1,17 @@
 import requests
 import base64
 #from typing import List, Dict
-import json
+#import json
 from time import sleep
 import pandas as pd
 import os
 from crenciais import Credential
 import multiprocessing
-from datetime import datetime
-import random
-import string
+#from datetime import datetime
+#import random
+#import string
+from typing import List
+from tratarDados import Tratar
 
 class Consume:
     @property
@@ -80,6 +82,9 @@ class Consume:
         else:
             raise Exception(f"erro ao consumir api, {response.status_code=}; {response.content}")
     
+    def request_api_multi(self, queue:multiprocessing.Queue,url:str):
+        queue.put(self.request_api(url))
+    
     def all_pages(self, url:str) -> list:# -> requests.models.Response:
         first_response:dict = self.request_api(url).json()
         keys:list = list(first_response.keys())
@@ -98,12 +103,8 @@ class Consume:
         
         return content
         
-        # if self.next_page != None:
-        #     return self.request_api(self.next_page)
-        # else:
-        #     return None
     
-    def all_tickets(self, *, url_patter:str, contador:int=0) -> dict|list:
+    def all_tickets(self, *, url_patter:str, contador:int=0) -> str|dict|list:
         while url_patter.endswith('/'):
             url_patter = url_patter[:-1]
             
@@ -111,44 +112,50 @@ class Consume:
         if not os.path.exists(path_temp):
             os.makedirs(path_temp)
         
-        file_name_temp:str = path_temp + f"File_temp_{str(datetime.now()).replace(':', '_')}_.json"
-        pd.DataFrame().to_json(file_name_temp, orient='records')
+        file_name_temp:str = path_temp + "tickets.json"
+        if os.path.exists(file_name_temp):
+            os.unlink(file_name_temp)
         content_variable:list = []
         
-        
         while True:
-            list_ids:list = [str((num+1)+contador) for num in range(100)]
-        
-            url:str = f"{url_patter}/api/v2/tickets/show_many.json?ids="
+            quant = 6
+            threads:List[dict] = [{"fila":multiprocessing.Queue(),"processo": None} for _ in range(quant)]
+            for thread in threads:
+                list_ids:list = [str((num+1)+contador) for num in range(100)]
             
-            url_mod:str = url + ','.join(list_ids)
-            print(url + list_ids[0] + ", " + list_ids[-1])
+                url:str = f"{url_patter}/api/v2/tickets/show_many.json?ids="
+                
+                url_mod:str = url + ','.join(list_ids)
+                print(url + list_ids[0] + ", " + list_ids[-1])
+                
+                thread["processo"] = multiprocessing.Process(target=self.request_api_multi, args=(thread["fila"],url_mod))
+                thread["processo"].start()
+                
+                contador += 100
             
-            api:dict = self.request_api(url_mod).json()
-            content:dict|list = api.get('tickets')# type: ignore
-            
-            #final_content += content
-            content_variable += content
-            if (len(content_variable) >= 5000) or (not content):
+            for thread in threads:
+                temp_fila:multiprocessing.Queue = thread['fila']
+                content:dict|list = temp_fila.get().json().get('tickets')# type: ignore
+                content_variable += content
+                
+             #final_content += content    
+            if (len(content_variable) >= 5000) or (not content_variable):
                 #print("alimentou")
-                #df_temp = pd.read_json(file_name_temp)
-                #acumulate_temp_df = pd.DataFrame(content_variable)
-                pd.DataFrame(content_variable).to_json(file_name_temp, orient='records', lines=True, mode='a')
-                #pd.concat([df_temp, acumulate_temp_df], ignore_index=True).to_json(file_name_temp, orient='records')
-                #del df_temp
-                #del acumulate_temp_df
-                content_variable = []             
+                df = pd.DataFrame(content_variable)
+                df = Tratar.start(df)
+                df.to_json(file_name_temp, orient='records', lines=True, mode='a')
+                content_variable = []  
             
             if not content:
                 break
             
-            contador += 100
+            
         
 
-        final_content:dict = pd.read_json(file_name_temp, orient='records', lines=True).to_dict()
+        #final_content:dict = pd.read_json(file_name_temp, orient='records').to_dict()
 
-        os.unlink(file_name_temp)
-        return final_content
+        #os.unlink(file_name_temp)
+        return file_name_temp
